# api/public/doc/v2/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Makeries API documentation",
    "version": "1.0.0",
    "contact": {
      "email": "apiteam@makeries.co.uk"
    }
  },
  "basePath": "/v2",
  "schemes": [
    "http"
  ],
  "securityDefinitions": {
    "apiKey": {
      "type": "apiKey",
      "in": "header",
      "name": "token"
    }
  },
  "paths": {
    "/maker": {
      "get": {
        "tags": [
          "maker"
        ],
        "summary": "Return makeries",
        "description": "Return all makeries",
        "operationId": "getMakers",
        "produces": [
          "application/json"
        ],
        "security": [
          {
            "apiKey": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "name",
            "required": false,
            "type": "string"
          },
          {
            "in": "query",
            "name": "address",
            "required": false,
            "type": "string"
          },
          {
            "in": "query",
            "name": "city",
            "required": false,
            "type": "string"
          },
          {
            "in": "query",
            "name": "postcode",
            "required": false,
            "type": "string"
          },
          {
            "in": "query",
            "name": "email",
            "required": false,
            "type": "string"
          },
          {
            "in": "query",
            "name": "telephone",
            "required": false,
            "type": "string"
          },
          {
            "in": "query",
            "name": "website",
            "required": false,
            "type": "string"
          },
          {
            "in": "query",
            "name": "social1",
            "required": false,
            "type": "string"
          },
          {
            "in": "query",
            "name": "social2",
            "required": false,
            "type": "string"
          },
          {
            "in": "query",
            "name": "social3",
            "required": false,
            "type": "string"
          },
          {
            "in": "query",
            "name": "mapUrl",
            "required": false,
            "type": "string"
          },
          {
            "in": "query",
            "name": "adminEmail",
            "required": false,
            "type": "string"
          },
          {
            "in": "query",
            "name": "briefDescription",
            "required": false,
            "type": "string"
          },
          {
            "in": "query",
            "name": "longDescription",
            "required": false,
            "type": "string"
          },
          {
            "in": "query",
            "name": "region.id",
            "required": false,
            "type": "string"
          },
          {
            "in": "query",
            "name": "published",
            "required": false,
            "type": "string",
            "enum": [
              "yes",
              "no"
            ]
          },
          {
            "in": "query",
            "name": "featured",
            "required": false,
            "type": "string",
            "enum": [
              "yes",
              "no"
            ]
          },
          {
            "in": "query",
            "name": "subscription",
            "required": false,
            "type": "string",
            "enum": [
              "yes",
              "no"
            ]
          },
          {
            "in": "query",
            "name": "sort_by",
            "required": false,
            "type": "string"
          },
          {
            "in": "query",
            "name": "page",
            "required": false,
            "type": "integer"
          },
          {
            "in": "query",
            "name": "per_page",
            "required": false,
            "type": "integer"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "items": {
                "$ref": "#/definitions/Maker"
              }
            }
          },
          "400": {
            "description": "Invalid data supplied"
          }
        }
      },
      "post": {
        "tags": [
          "maker"
        ],
        "summary": "Create new maker",
        "description": "Create a maker",
        "operationId": "createMaker",
        "produces": [
          "application/json"
        ],
        "security": [
          {
            "apiKey": []
          }
        ],
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "maker",
            "description": "The maker to create.",
            "schema": {
              "$ref": "#/components/schemas/Maker"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "$ref": "#/definitions/Maker"
            }
          },
          "400": {
            "description": "Invalid data supplied"
          }
        }
      }
    },
    "/maker/{id}": {
      "get": {
        "tags": [
          "maker"
        ],
        "summary": "Return a maker",
        "description": "Return one maker by ID",
        "operationId": "getMaker",
        "produces": [
          "application/json"
        ],
        "security": [
          {
            "apiKey": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID of maker to return",
            "required": true,
            "type": "integer"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "$ref": "#/definitions/Maker"
            }
          },
          "400": {
            "description": "Invalid data supplied"
          }
        }
      },
      "put": {
        "tags": [
          "maker"
        ],
        "summary": "Update maker",
        "description": "Update maker detail by ID",
        "operationId": "updateMaker",
        "produces": [
          "application/json"
        ],
        "security": [
          {
            "apiKey": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID of maker to update",
            "required": true,
            "type": "integer"
          },
          {
            "in": "body",
            "name": "maker",
            "description": "The maker to update.",
            "schema": {
              "$ref": "#/components/schemas/Maker"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "$ref": "#/definitions/Maker"
            }
          },
          "400": {
            "description": "Invalid data supplied"
          }
        }
      },
      "delete": {
        "tags": [
          "maker"
        ],
        "summary": "Delete maker",
        "description": "Delete maker by ID",
        "operationId": "deleteMaker",
        "produces": [
          "application/json"
        ],
        "security": [
          {
            "apiKey": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID of maker to delete",
            "required": true,
            "type": "integer"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation"
          },
          "400": {
            "description": "Invalid data supplied"
          }
        }
      }
    },
    "/maker/{id}/image": {
      "get": {
        "tags": [
          "maker"
        ],
        "summary": "Get maker images",
        "description": "get list of images for maker",
        "operationId": "getMakerImages",
        "produces": [
          "application/json"
        ],
        "security": [
          {
            "apiKey": []
          }
        ],
        "consumes": [
          "application/x-www-form-urlencoded"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID of maker",
            "required": true,
            "type": "integer"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "$ref": "#/definitions/Image"
            }
          },
          "400": {
            "description": "Invalid ID supplied"
          },
          "404": {
            "description": "Image not found"
          }
        }
      },
      "post": {
        "tags": [
          "maker"
        ],
        "summary": "Add image to maker",
        "description": "add image maker",
        "operationId": "addMakerImage",
        "produces": [
          "application/json"
        ],
        "security": [
          {
            "apiKey": []
          }
        ],
        "consumes": [
          "multipart/form-data"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID of maker",
            "required": true,
            "type": "integer"
          },
          {
            "name": "name",
            "in": "formData",
            "description": "name of file",
            "required": true,
            "type": "string"
          },
          {
            "in": "formData",
            "name": "upfile",
            "type": "file",
            "description": "The file to upload.",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "$ref": "#/definitions/Image"
            }
          },
          "400": {
            "description": "Invalid ID supplied"
          },
          "404": {
            "description": "Maker not found"
          }
        }
      }
    },
    "/maker/{id}/image/{imgId}": {
      "get": {
        "tags": [
          "maker"
        ],
        "summary": "Download image",
        "description": "Download image",
        "operationId": "getMakerImage",
        "produces": [
          "application/json"
        ],
        "security": [
          {
            "apiKey": []
          }
        ],
        "consumes": [
          "application/x-www-form-urlencoded"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID of maker",
            "required": true,
            "type": "integer"
          },
          {
            "name": "imgId",
            "in": "path",
            "description": "ID of image to download",
            "required": true,
            "type": "integer"
          }
        ],
        "responses": {
          "200": {
            "description": "image data"
          },
          "204": {
            "description": "successful operation"
          }
        }
      },
      "delete": {
        "tags": [
          "maker"
        ],
        "summary": "Delete image",
        "description": "Delete image",
        "operationId": "deleteMakerImage",
        "produces": [
          "application/json"
        ],
        "security": [
          {
            "apiKey": []
          }
        ],
        "consumes": [
          "application/x-www-form-urlencoded"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID of maker",
            "required": true,
            "type": "integer"
          },
          {
            "name": "imgId",
            "in": "path",
            "description": "ID of image to delete",
            "required": true,
            "type": "integer"
          }
        ],
        "responses": {
          "204": {
            "description": "successful operation"
          }
        }
      }
    },
...
    "/project/{id}/image": {
      "get": {
        "tags": [
          "project"
        ],
        "summary": "Get project images",
        "description": "get list of images for project",
        "operationId": "getProjectImages",
        "produces": [
          "application/json"
        ],
        "security": [
          {
            "apiKey": []
          }
        ],
        "consumes": [
          "application/x-www-form-urlencoded"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID of project",
            "required": true,
            "type": "integer"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "$ref": "#/definitions/Image"
            }
          },
          "400": {
            "description": "Invalid ID supplied"
          },
          "404": {
            "description": "Image not found"
          }
        }
      },
      "post": {
        "tags": [
          "project"
        ],
        "summary": "Add image to project",
        "description": "add image project",
        "operationId": "addProjectImage",
        "produces": [
          "application/json"
        ],
        "security": [
          {
            "apiKey": []
          }
        ],
        "consumes": [
          "multipart/form-data"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID of project",
            "required": true,
            "type": "integer"
          },
          {
            "name": "name",
            "in": "formData",
            "description": "name of file",
            "required": true,
            "type": "string"
          },
          {
            "in": "formData",
            "name": "upfile",
            "type": "file",
            "description": "The file to upload.",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "$ref": "#/definitions/Image"
            }
          },
          "400": {
            "description": "Invalid ID supplied"
          },
          "404": {
            "description": "Project not found"
          }
        }
      }
    },
    "/project/{id}/image/{imgId}": {
      "get": {
        "tags": [
          "project"
        ],
        "summary": "Download image",
        "description": "Download image",
        "operationId": "getProjectImage",
        "produces": [
          "application/json"
        ],
        "security": [
          {
            "apiKey": []
          }
        ],
        "consumes": [
          "application/x-www-form-urlencoded"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID of project",
            "required": true,
            "type": "integer"
          },
          {
            "name": "imgId",
            "in": "path",
            "description": "ID of image to download",
            "required": true,
            "type": "integer"
          }
        ],
        "responses": {
          "200": {
            "description": "image data"
          },
          "204": {
            "description": "successful operation"
          }
        }
      },
      "delete": {
        "tags": [
          "project"
        ],
        "summary": "Delete image",
        "description": "Delete image",
        "operationId": "deleteProjectImage",
        "produces": [
          "application/json"
        ],
        "security": [
          {
            "apiKey": []
          }
        ],
        "consumes": [
          "application/x-www-form-urlencoded"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID of project",
            "required": true,
            "type": "integer"
          },
          {
            "name": "imgId",
            "in": "path",
            "description": "ID of image to delete",
            "required": true,
            "type": "integer"
          }
        ],
        "responses": {
          "204": {
            "description": "successful operation"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Maker": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "address1": {
            "type": "string"
          },
          "address2": {
            "type": "string"
          },
          "city": {
            "type": "string"
          },
          "postcode": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "telephone": {
            "type": "string"
          },
          "website": {
            "type": "string"
          },
          "social1": {
            "type": "string"
          },
          "social2": {
            "type": "string"
          },
          "social3": {
            "type": "string"
          },
          "map_url": {
            "type": "string"
          },
          "admin_email": {
            "type": "string"
          },
          "brief_description": {
            "type": "string"
          },
          "long_description": {
            "type": "string"
          },
          "regions": {
            "type": "array",
            "items": {
              "type": "integer"
            }
          },
          "published": {
            "type": "string",
            "enum": [
              "yes",
              "no"
            ]
          },
          "featured": {
            "type": "string",
            "enum": [
              "yes",
              "no"
            ]
          },
          "subscription": {
            "type": "string",
            "enum": [
              "yes",
              "no"
            ]
          },
          "business_types": {
            "type": "array",
            "items": {
              "type": "integer"
            }
          },
          "service_types": {
            "type": "array",
            "items": {
              "type": "integer"
            }
          },
          "products": {
            "type": "array",
            "items": {
              "type": "integer"
            }
          },
          "medias": {
            "type": "array",
            "items": {
              "type": "integer"
            }
          },
          "projects": {
            "type": "array",
            "items": {
              "type": "integer"
            }
          },
          "maker_groups": {
            "type": "array",
            "items": {
              "type": "integer"
            }
          }
        }
      }
    }
  },
  "definitions": {
    "Maker": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer"
        },
        "name": {
          "type": "string"
        },
        "address1": {
          "type": "string"
        },
        "address2": {
          "type": "string"
        },
        "city": {
          "type": "string"
        },
        "postcode": {
          "type": "string"
        },
        "email": {
          "type": "string"
        },
        "telephone": {
          "type": "string"
        },
        "website": {
          "type": "string"
        },
        "social1": {
          "type": "string"
        },
        "social2": {
          "type": "string"
        },
        "social3": {
          "type": "string"
        },
        "mapUrl": {
          "type": "string"
        },
        "admin_email": {
          "type": "string"
        },
        "brief_description": {
          "type": "string"
        },
        "long_description": {
          "type": "string"
        },
        "region": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Region"
          }
        },
        "published": {
          "type": "string",
          "enum": [
            "yes",
            "no"
          ]
        },
        "featured": {
          "type": "string",
          "enum": [
            "yes",
            "no"
          ]
        },
        "subscription": {
          "type": "string",
          "enum": [
            "yes",
            "no"
          ]
        },
        "businesstypes": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/BusinessType"
          }
        },
        "servicetypes": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/ServiceType"
          }
        },
        "products": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Product"
          }
        },
        "medias": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Media"
          }
        },
        "projects": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Project"
          }
        },
        "makergroups": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Product"
          }
        },
        "images": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Image"
          }
        }
      }
    },
    "Image": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer"
        },
        "name": {
          "type": "string"
        },
        "image_data": {
          "type": "string"
        }
      }
    },
    "BusinessType": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer"
        },
        "name": {
          "type": "string"
        }
      }
    },
...
    "ServiceType": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer"
        },
        "name": {
          "type": "string"
        }
      }
    },
    "Media": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer"
        },
        "name": {
          "type": "string"
        },
        "image_data": {
          "type": "string"
        },
        "url": {
          "type": "string"
        }
      }
    },
    "Project": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer"
        },
        "maker)id": {
          "type": "integer"
        },
        "name": {
          "type": "string"
        },
        "description": {
          "type": "string"
        },
        "images": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Image"
          }
        }
      }
    },
    "Product": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer"
        },
        "name": {
          "type": "string"
        }
      }
    },
    "Region": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer"
        },
        "name": {
          "type": "string"
        }
      }
    }
  },
...
}